Move path helper method to PathUtils and fix namespace
